Fix warning

+++ SGSLineGraph/SGSLineGraphView.m
@@ -129,7 +129,7 @@
     UIGraphicsPushContext(ctx);
     CGContextSetRGBFillColor(ctx, 0.2f, 0.2f, 0.2f, 1.0f);
     
-    int n_div;
+    unsigned long n_div;
     float scale_min, scale_max, div_height;
     float top_margin = 35;
     float bottom_margin = 25;
@@ -230,7 +230,7 @@
 - (void) setupGraphPaths {
     
     
-    int n_div;
+    unsigned long n_div;
     float scale_min, scale_max, div_height;
     float top_margin = 35;
     float bottom_margin = 25;
